[Refactor] 영속 데이터 네임스페이스 분리
