+++ RAW_unity/Assets/_RAW/Scripts/Network/Session/ServerPlayerSessionManager.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
+using RAW.Persistence;
 using UnityEngine;
 
 namespace RAW.Network
