+++ RAW_unity/Assets/_RAW/Scripts/Persistence/InMemoryPlayerDataRepository.cs
@@ -4,7 +4,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using UnityEngine;
 
-namespace RAW.Network
+namespace RAW.Persistence
 {
 	public sealed class InMemoryPlayerDataRepository : PlayerDataRepository
 	{
